Refactor model folder structure.
