+++ videollama2/model/encoder.py
@@ -1,3 +1,5 @@
+import os
+
 import torch
 import torch.nn as nn
 
@@ -82,3 +84,15 @@ class CLIPVisionTower(nn.Module):
     @property
     def num_patches_per_side(self):
         return self.config.image_size // self.config.patch_size
+
+
+def build_vision_tower(vision_tower_cfg, **kwargs):
+    vision_tower = getattr(vision_tower_cfg, 'mm_vision_tower', getattr(vision_tower_cfg, 'vision_tower', None))
+
+    is_absolute_path_exists = os.path.exists(vision_tower)
+    if  vision_tower.startswith("openai") or vision_tower.startswith("laion") or 'clip' in vision_tower:
+        vision_tower = CLIPVisionTower(vision_tower, args=vision_tower_cfg, **kwargs)
+    else:
+        raise ValueError(f'Unknown vision tower: {vision_tower}')
+
+    return vision_tower
