--- videollama2/model/multimodal_projector/__init__.py
@@ -1,1 +1,0 @@
-from .builder import load_mm_projector
