+++ videollama2/model/videollama2_llama.py
@@ -1,3 +1,4 @@
+# Adopted from: https://github.com/haotian-liu/LLaVA. Below is the original copyright:
 #    Copyright 2023 Haotian Liu
 #
 #    Licensed under the Apache License, Version 2.0 (the "License");
@@ -17,35 +18,33 @@ from typing import List, Optional, Tuple, Union
 
 import torch
 import torch.nn as nn
-from torch.nn import CrossEntropyLoss
 
 from transformers import AutoConfig, AutoModelForCausalLM, \
-                         MixtralConfig, MixtralModel, MixtralForCausalLM
-
+                         LlamaConfig, LlamaModel, LlamaForCausalLM
 from transformers.modeling_outputs import CausalLMOutputWithPast
 from transformers.generation.utils import GenerateOutput
 
-from ..videollama2_arch import Videollama2MetaModel, Videollama2MetaForCausalLM
+from .videollama2_arch import Videollama2MetaModel, Videollama2MetaForCausalLM
 
 
-class Videollama2MixtralConfig(MixtralConfig):
-    model_type = "videollama2_mixtral"
+class Videollama2Config(LlamaConfig):
+    model_type = "videollama2_llama"
 
 
-class Videollama2MixtralModel(Videollama2MetaModel, MixtralModel):
-    config_class = Videollama2MixtralConfig
+class Videollama2LlamaModel(Videollama2MetaModel, LlamaModel):
+    config_class = Videollama2Config
 
-    def __init__(self, config: MixtralConfig):
-        super(Videollama2MixtralModel, self).__init__(config)
+    def __init__(self, config: LlamaConfig):
+        super(Videollama2LlamaModel, self).__init__(config)
 
 
-class Videollama2MixtralForCausalLM(MixtralForCausalLM, Videollama2MetaForCausalLM):
-    config_class = Videollama2MixtralConfig
+class Videollama2LlamaForCausalLM(LlamaForCausalLM, Videollama2MetaForCausalLM):
+    config_class = Videollama2Config
 
     def __init__(self, config, **kwargs):
-        super(MixtralForCausalLM, self).__init__(config)
-        self.model = Videollama2MixtralModel(config)
-        # self.pretraining_tp = config.pretraining_tp
+        super(LlamaForCausalLM, self).__init__(config)
+        self.model = Videollama2LlamaModel(config)
+        self.pretraining_tp = config.pretraining_tp
         self.vocab_size = config.vocab_size
         self.lm_head = nn.Linear(config.hidden_size, config.vocab_size, bias=False)
 
@@ -86,7 +85,7 @@ class Videollama2MixtralForCausalLM(MixtralForCausalLM, Videollama2MetaForCausal
                 images
             )
 
-        return super().forward(
+        outputs = super().forward(
             input_ids=input_ids,
             attention_mask=attention_mask,
             past_key_values=past_key_values,
@@ -98,13 +97,15 @@ class Videollama2MixtralForCausalLM(MixtralForCausalLM, Videollama2MetaForCausal
             return_dict=return_dict
         )
 
+        outputs.labels = labels
+
+        return outputs
+
     @torch.no_grad()
     def generate(
         self,
         inputs: Optional[torch.Tensor] = None,
-        images_or_videos: Optional[torch.Tensor] = None,
-        timestamps: Optional[torch.Tensor] = None,
-        modal_list: Optional[torch.Tensor] = None,
+        images: Optional[torch.Tensor] = None,
         **kwargs,
     ) -> Union[GenerateOutput, torch.LongTensor]:
         position_ids = kwargs.pop("position_ids", None)
@@ -112,8 +113,7 @@ class Videollama2MixtralForCausalLM(MixtralForCausalLM, Videollama2MetaForCausal
         if "inputs_embeds" in kwargs:
             raise NotImplementedError("`inputs_embeds` is not supported")
 
-        if images_or_videos is not None:
-            X_modalities = [images_or_videos, modal_list] if timestamps is None else [images_or_videos, modal_list, timestamps]
+        if images is not None:
             (
                 input_ids,
                 attention_mask,
@@ -125,7 +125,7 @@ class Videollama2MixtralForCausalLM(MixtralForCausalLM, Videollama2MetaForCausal
                 attention_mask=attention_mask,
                 past_key_values=None,
                 labels=None,
-                X_modalities=X_modalities
+                images=images
             )
         else:
             inputs_embeds = self.get_model().embed_tokens(inputs)
@@ -146,5 +146,6 @@ class Videollama2MixtralForCausalLM(MixtralForCausalLM, Videollama2MetaForCausal
             _inputs['images'] = images
         return _inputs
 
-AutoConfig.register("videollama2_mixtral", Videollama2MixtralConfig)
-AutoModelForCausalLM.register(Videollama2MixtralConfig, Videollama2MixtralForCausalLM)
+
+AutoConfig.register("videollama2_llama", Videollama2Config)
+AutoModelForCausalLM.register(Videollama2Config, Videollama2LlamaForCausalLM)
