+++ videollama2/model/videollama2_mistral.py
@@ -18,32 +18,34 @@ from typing import List, Optional, Tuple, Union
 
 import torch
 import torch.nn as nn
+from torch.nn import CrossEntropyLoss
 
 from transformers import AutoConfig, AutoModelForCausalLM, \
-                         Qwen2Config, Qwen2Model, Qwen2ForCausalLM
+                         MistralConfig, MistralModel, MistralForCausalLM
+
 from transformers.modeling_outputs import CausalLMOutputWithPast
 from transformers.generation.utils import GenerateOutput
 
-from ..videollama2_arch import Videollama2MetaModel, Videollama2MetaForCausalLM
+from .videollama2_arch import Videollama2MetaModel, Videollama2MetaForCausalLM
 
 
-class Videollama2Qwen2Config(Qwen2Config):
-    model_type = "videollama2_qwen2"
+class Videollama2MistralConfig(MistralConfig):
+    model_type = "videollama2_mistral"
 
 
-class Videollama2Qwen2Model(Videollama2MetaModel, Qwen2Model):
-    config_class = Videollama2Qwen2Config
+class Videollama2MistralModel(Videollama2MetaModel, MistralModel):
+    config_class = Videollama2MistralConfig
 
-    def __init__(self, config: Videollama2Qwen2Config):
-        super(Videollama2Qwen2Model, self).__init__(config)
+    def __init__(self, config: MistralConfig):
+        super(Videollama2MistralModel, self).__init__(config)
 
 
-class Videollama2Qwen2ForCausalLM(Qwen2ForCausalLM, Videollama2MetaForCausalLM):
-    config_class = Videollama2Qwen2Config
+class Videollama2MistralForCausalLM(MistralForCausalLM, Videollama2MetaForCausalLM):
+    config_class = Videollama2MistralConfig
 
     def __init__(self, config, **kwargs):
-        super(Qwen2ForCausalLM, self).__init__(config)
-        self.model = Videollama2Qwen2Model(config)
+        super(MistralForCausalLM, self).__init__(config)
+        self.model = Videollama2MistralModel(config)
         # self.pretraining_tp = config.pretraining_tp
         self.vocab_size = config.vocab_size
         self.lm_head = nn.Linear(config.hidden_size, config.vocab_size, bias=False)
@@ -85,7 +87,7 @@ class Videollama2Qwen2ForCausalLM(Qwen2ForCausalLM, Videollama2MetaForCausalLM):
                 images
             )
 
-        return super().forward(
+        outputs = super().forward(
             input_ids=input_ids,
             attention_mask=attention_mask,
             past_key_values=past_key_values,
@@ -97,12 +99,15 @@ class Videollama2Qwen2ForCausalLM(Qwen2ForCausalLM, Videollama2MetaForCausalLM):
             return_dict=return_dict
         )
 
+        outputs.labels = labels
+
+        return outputs
+
     @torch.no_grad()
     def generate(
         self,
         inputs: Optional[torch.Tensor] = None,
-        images_or_videos: Optional[torch.Tensor] = None,
-        modal_list: Optional[torch.Tensor] = None,
+        images: Optional[torch.Tensor] = None,
         **kwargs,
     ) -> Union[GenerateOutput, torch.LongTensor]:
         position_ids = kwargs.pop("position_ids", None)
@@ -110,7 +115,7 @@ class Videollama2Qwen2ForCausalLM(Qwen2ForCausalLM, Videollama2MetaForCausalLM):
         if "inputs_embeds" in kwargs:
             raise NotImplementedError("`inputs_embeds` is not supported")
 
-        if images_or_videos is not None:
+        if images is not None:
             (
                 input_ids,
                 attention_mask,
@@ -122,7 +127,7 @@ class Videollama2Qwen2ForCausalLM(Qwen2ForCausalLM, Videollama2MetaForCausalLM):
                 attention_mask=attention_mask,
                 past_key_values=None,
                 labels=None,
-                X_modalities=[images_or_videos, modal_list]
+                images=images
             )
         else:
             inputs_embeds = self.get_model().embed_tokens(inputs)
@@ -144,5 +149,5 @@ class Videollama2Qwen2ForCausalLM(Qwen2ForCausalLM, Videollama2MetaForCausalLM):
         return _inputs
 
 
-AutoConfig.register("videollama2_qwen2", Videollama2Qwen2Config)
-AutoModelForCausalLM.register(Videollama2Qwen2Config, Videollama2Qwen2ForCausalLM)
+AutoConfig.register("videollama2_mistral", Videollama2MistralConfig)
+AutoModelForCausalLM.register(Videollama2MistralConfig, Videollama2MistralForCausalLM)
