+++ videollama2/model/videollama2_mixtral.py
@@ -1,4 +1,3 @@
-# Adopted from: https://github.com/haotian-liu/LLaVA. Below is the original copyright:
 #    Copyright 2023 Haotian Liu
 #
 #    Licensed under the Apache License, Version 2.0 (the "License");
@@ -21,31 +20,31 @@ import torch.nn as nn
 from torch.nn import CrossEntropyLoss
 
 from transformers import AutoConfig, AutoModelForCausalLM, \
-                         MistralConfig, MistralModel, MistralForCausalLM
+                         MixtralConfig, MixtralModel, MixtralForCausalLM
 
 from transformers.modeling_outputs import CausalLMOutputWithPast
 from transformers.generation.utils import GenerateOutput
 
-from ..videollama2_arch import Videollama2MetaModel, Videollama2MetaForCausalLM
+from .videollama2_arch import Videollama2MetaModel, Videollama2MetaForCausalLM
 
 
-class Videollama2MistralConfig(MistralConfig):
-    model_type = "videollama2_mistral"
+class Videollama2MixtralConfig(MixtralConfig):
+    model_type = "videollama2_mixtral"
 
 
-class Videollama2MistralModel(Videollama2MetaModel, MistralModel):
-    config_class = Videollama2MistralConfig
+class Videollama2MixtralModel(Videollama2MetaModel, MixtralModel):
+    config_class = Videollama2MixtralConfig
 
-    def __init__(self, config: MistralConfig):
-        super(Videollama2MistralModel, self).__init__(config)
+    def __init__(self, config: MixtralConfig):
+        super(Videollama2MixtralModel, self).__init__(config)
 
 
-class Videollama2MistralForCausalLM(MistralForCausalLM, Videollama2MetaForCausalLM):
-    config_class = Videollama2MistralConfig
+class Videollama2MixtralForCausalLM(MixtralForCausalLM, Videollama2MetaForCausalLM):
+    config_class = Videollama2MixtralConfig
 
     def __init__(self, config, **kwargs):
-        super(MistralForCausalLM, self).__init__(config)
-        self.model = Videollama2MistralModel(config)
+        super(MixtralForCausalLM, self).__init__(config)
+        self.model = Videollama2MixtralModel(config)
         # self.pretraining_tp = config.pretraining_tp
         self.vocab_size = config.vocab_size
         self.lm_head = nn.Linear(config.hidden_size, config.vocab_size, bias=False)
@@ -103,8 +102,7 @@ class Videollama2MistralForCausalLM(MistralForCausalLM, Videollama2MetaForCausal
     def generate(
         self,
         inputs: Optional[torch.Tensor] = None,
-        images_or_videos: Optional[torch.Tensor] = None,
-        modal_list: Optional[torch.Tensor] = None,
+        images: Optional[torch.Tensor] = None,
         **kwargs,
     ) -> Union[GenerateOutput, torch.LongTensor]:
         position_ids = kwargs.pop("position_ids", None)
@@ -112,7 +110,7 @@ class Videollama2MistralForCausalLM(MistralForCausalLM, Videollama2MetaForCausal
         if "inputs_embeds" in kwargs:
             raise NotImplementedError("`inputs_embeds` is not supported")
 
-        if images_or_videos is not None:
+        if images is not None:
             (
                 input_ids,
                 attention_mask,
@@ -124,7 +122,7 @@ class Videollama2MistralForCausalLM(MistralForCausalLM, Videollama2MetaForCausal
                 attention_mask=attention_mask,
                 past_key_values=None,
                 labels=None,
-                X_modalities=[images_or_videos, modal_list]
+                images=images
             )
         else:
             inputs_embeds = self.get_model().embed_tokens(inputs)
@@ -146,5 +144,5 @@ class Videollama2MistralForCausalLM(MistralForCausalLM, Videollama2MetaForCausal
         return _inputs
 
 
-AutoConfig.register("videollama2_mistral", Videollama2MistralConfig)
-AutoModelForCausalLM.register(Videollama2MistralConfig, Videollama2MistralForCausalLM)
+AutoConfig.register("videollama2_mixtral", Videollama2MixtralConfig)
+AutoModelForCausalLM.register(Videollama2MixtralConfig, Videollama2MixtralForCausalLM)
